+++ lib/src/data_transfer_objects/wallet/list_transactions/list_transactions_params.freezed.dart
@@ -15,7 +15,7 @@ T _$identity<T>(T value) => value;
 /// @nodoc
 mixin _$ListTransactionsParams {
 
-@JsonKey(name: 'asset') String? get asset;@JsonKey(name: 'min_topoheight') int? get minTopoHeight;@JsonKey(name: 'max_topoheight') int? get maxTopoHeight;@JsonKey(name: 'min_timestamp') int? get minTimestamp;@JsonKey(name: 'max_timestamp') int? get maxTimestamp;@JsonKey(name: 'address') String? get address;@JsonKey(name: 'accept_incoming') bool? get acceptIncoming;@JsonKey(name: 'accept_outgoing') bool? get acceptOutgoing;@JsonKey(name: 'accept_coinbase') bool? get acceptCoinbase;@JsonKey(name: 'accept_burn') bool? get acceptBurn;@JsonKey(name: 'accept_blob') bool? get acceptBlob;// TODO
+@JsonKey(name: 'asset') String? get asset;@JsonKey(name: 'min_topoheight') int? get minTopoHeight;@JsonKey(name: 'max_topoheight') int? get maxTopoHeight;@JsonKey(name: 'min_timestamp') int? get minTimestamp;@JsonKey(name: 'max_timestamp') int? get maxTimestamp;@JsonKey(name: 'address') String? get address;@JsonKey(name: 'contract') String? get contract;@JsonKey(name: 'accept_incoming') bool? get acceptIncoming;@JsonKey(name: 'accept_outgoing') bool? get acceptOutgoing;@JsonKey(name: 'accept_coinbase') bool? get acceptCoinbase;@JsonKey(name: 'accept_burn') bool? get acceptBurn;@JsonKey(name: 'accept_blob') bool? get acceptBlob;// TODO
 @JsonKey(name: 'query') dynamic get query;@JsonKey(name: 'limit') int? get limit;@JsonKey(name: 'skip') int? get skip;
 /// Create a copy of ListTransactionsParams
 /// with the given fields replaced by the non-null parameter values.
@@ -29,16 +29,16 @@ $ListTransactionsParamsCopyWith<ListTransactionsParams> get copyWith => _$ListTr
 
 @override
 bool operator ==(Object other) {
-  return identical(this, other) || (other.runtimeType == runtimeType&&other is ListTransactionsParams&&(identical(other.asset, asset) || other.asset == asset)&&(identical(other.minTopoHeight, minTopoHeight) || other.minTopoHeight == minTopoHeight)&&(identical(other.maxTopoHeight, maxTopoHeight) || other.maxTopoHeight == maxTopoHeight)&&(identical(other.minTimestamp, minTimestamp) || other.minTimestamp == minTimestamp)&&(identical(other.maxTimestamp, maxTimestamp) || other.maxTimestamp == maxTimestamp)&&(identical(other.address, address) || other.address == address)&&(identical(other.acceptIncoming, acceptIncoming) || other.acceptIncoming == acceptIncoming)&&(identical(other.acceptOutgoing, acceptOutgoing) || other.acceptOutgoing == acceptOutgoing)&&(identical(other.acceptCoinbase, acceptCoinbase) || other.acceptCoinbase == acceptCoinbase)&&(identical(other.acceptBurn, acceptBurn) || other.acceptBurn == acceptBurn)&&(identical(other.acceptBlob, acceptBlob) || other.acceptBlob == acceptBlob)&&const DeepCollectionEquality().equals(other.query, query)&&(identical(other.limit, limit) || other.limit == limit)&&(identical(other.skip, skip) || other.skip == skip));
+  return identical(this, other) || (other.runtimeType == runtimeType&&other is ListTransactionsParams&&(identical(other.asset, asset) || other.asset == asset)&&(identical(other.minTopoHeight, minTopoHeight) || other.minTopoHeight == minTopoHeight)&&(identical(other.maxTopoHeight, maxTopoHeight) || other.maxTopoHeight == maxTopoHeight)&&(identical(other.minTimestamp, minTimestamp) || other.minTimestamp == minTimestamp)&&(identical(other.maxTimestamp, maxTimestamp) || other.maxTimestamp == maxTimestamp)&&(identical(other.address, address) || other.address == address)&&(identical(other.contract, contract) || other.contract == contract)&&(identical(other.acceptIncoming, acceptIncoming) || other.acceptIncoming == acceptIncoming)&&(identical(other.acceptOutgoing, acceptOutgoing) || other.acceptOutgoing == acceptOutgoing)&&(identical(other.acceptCoinbase, acceptCoinbase) || other.acceptCoinbase == acceptCoinbase)&&(identical(other.acceptBurn, acceptBurn) || other.acceptBurn == acceptBurn)&&(identical(other.acceptBlob, acceptBlob) || other.acceptBlob == acceptBlob)&&const DeepCollectionEquality().equals(other.query, query)&&(identical(other.limit, limit) || other.limit == limit)&&(identical(other.skip, skip) || other.skip == skip));
 }
 
 @JsonKey(includeFromJson: false, includeToJson: false)
 @override
-int get hashCode => Object.hash(runtimeType,asset,minTopoHeight,maxTopoHeight,minTimestamp,maxTimestamp,address,acceptIncoming,acceptOutgoing,acceptCoinbase,acceptBurn,acceptBlob,const DeepCollectionEquality().hash(query),limit,skip);
+int get hashCode => Object.hash(runtimeType,asset,minTopoHeight,maxTopoHeight,minTimestamp,maxTimestamp,address,contract,acceptIncoming,acceptOutgoing,acceptCoinbase,acceptBurn,acceptBlob,const DeepCollectionEquality().hash(query),limit,skip);
 
 @override
 String toString() {
-  return 'ListTransactionsParams(asset: $asset, minTopoHeight: $minTopoHeight, maxTopoHeight: $maxTopoHeight, minTimestamp: $minTimestamp, maxTimestamp: $maxTimestamp, address: $address, acceptIncoming: $acceptIncoming, acceptOutgoing: $acceptOutgoing, acceptCoinbase: $acceptCoinbase, acceptBurn: $acceptBurn, acceptBlob: $acceptBlob, query: $query, limit: $limit, skip: $skip)';
+  return 'ListTransactionsParams(asset: $asset, minTopoHeight: $minTopoHeight, maxTopoHeight: $maxTopoHeight, minTimestamp: $minTimestamp, maxTimestamp: $maxTimestamp, address: $address, contract: $contract, acceptIncoming: $acceptIncoming, acceptOutgoing: $acceptOutgoing, acceptCoinbase: $acceptCoinbase, acceptBurn: $acceptBurn, acceptBlob: $acceptBlob, query: $query, limit: $limit, skip: $skip)';
 }
 
 
@@ -49,7 +49,7 @@ abstract mixin class $ListTransactionsParamsCopyWith<$Res>  {
   factory $ListTransactionsParamsCopyWith(ListTransactionsParams value, $Res Function(ListTransactionsParams) _then) = _$ListTransactionsParamsCopyWithImpl;
 @useResult
 $Res call({
-@JsonKey(name: 'asset') String? asset,@JsonKey(name: 'min_topoheight') int? minTopoHeight,@JsonKey(name: 'max_topoheight') int? maxTopoHeight,@JsonKey(name: 'min_timestamp') int? minTimestamp,@JsonKey(name: 'max_timestamp') int? maxTimestamp,@JsonKey(name: 'address') String? address,@JsonKey(name: 'accept_incoming') bool? acceptIncoming,@JsonKey(name: 'accept_outgoing') bool? acceptOutgoing,@JsonKey(name: 'accept_coinbase') bool? acceptCoinbase,@JsonKey(name: 'accept_burn') bool? acceptBurn,@JsonKey(name: 'accept_blob') bool? acceptBlob,@JsonKey(name: 'query') dynamic query,@JsonKey(name: 'limit') int? limit,@JsonKey(name: 'skip') int? skip
+@JsonKey(name: 'asset') String? asset,@JsonKey(name: 'min_topoheight') int? minTopoHeight,@JsonKey(name: 'max_topoheight') int? maxTopoHeight,@JsonKey(name: 'min_timestamp') int? minTimestamp,@JsonKey(name: 'max_timestamp') int? maxTimestamp,@JsonKey(name: 'address') String? address,@JsonKey(name: 'contract') String? contract,@JsonKey(name: 'accept_incoming') bool? acceptIncoming,@JsonKey(name: 'accept_outgoing') bool? acceptOutgoing,@JsonKey(name: 'accept_coinbase') bool? acceptCoinbase,@JsonKey(name: 'accept_burn') bool? acceptBurn,@JsonKey(name: 'accept_blob') bool? acceptBlob,@JsonKey(name: 'query') dynamic query,@JsonKey(name: 'limit') int? limit,@JsonKey(name: 'skip') int? skip
 });
 
 
@@ -66,7 +66,7 @@ class _$ListTransactionsParamsCopyWithImpl<$Res>
 
 /// Create a copy of ListTransactionsParams
 /// with the given fields replaced by the non-null parameter values.
-@pragma('vm:prefer-inline') @override $Res call({Object? asset = freezed,Object? minTopoHeight = freezed,Object? maxTopoHeight = freezed,Object? minTimestamp = freezed,Object? maxTimestamp = freezed,Object? address = freezed,Object? acceptIncoming = freezed,Object? acceptOutgoing = freezed,Object? acceptCoinbase = freezed,Object? acceptBurn = freezed,Object? acceptBlob = freezed,Object? query = freezed,Object? limit = freezed,Object? skip = freezed,}) {
+@pragma('vm:prefer-inline') @override $Res call({Object? asset = freezed,Object? minTopoHeight = freezed,Object? maxTopoHeight = freezed,Object? minTimestamp = freezed,Object? maxTimestamp = freezed,Object? address = freezed,Object? contract = freezed,Object? acceptIncoming = freezed,Object? acceptOutgoing = freezed,Object? acceptCoinbase = freezed,Object? acceptBurn = freezed,Object? acceptBlob = freezed,Object? query = freezed,Object? limit = freezed,Object? skip = freezed,}) {
   return _then(_self.copyWith(
 asset: freezed == asset ? _self.asset : asset // ignore: cast_nullable_to_non_nullable
 as String?,minTopoHeight: freezed == minTopoHeight ? _self.minTopoHeight : minTopoHeight // ignore: cast_nullable_to_non_nullable
@@ -74,6 +74,7 @@ as int?,maxTopoHeight: freezed == maxTopoHeight ? _self.maxTopoHeight : maxTopoH
 as int?,minTimestamp: freezed == minTimestamp ? _self.minTimestamp : minTimestamp // ignore: cast_nullable_to_non_nullable
 as int?,maxTimestamp: freezed == maxTimestamp ? _self.maxTimestamp : maxTimestamp // ignore: cast_nullable_to_non_nullable
 as int?,address: freezed == address ? _self.address : address // ignore: cast_nullable_to_non_nullable
+as String?,contract: freezed == contract ? _self.contract : contract // ignore: cast_nullable_to_non_nullable
 as String?,acceptIncoming: freezed == acceptIncoming ? _self.acceptIncoming : acceptIncoming // ignore: cast_nullable_to_non_nullable
 as bool?,acceptOutgoing: freezed == acceptOutgoing ? _self.acceptOutgoing : acceptOutgoing // ignore: cast_nullable_to_non_nullable
 as bool?,acceptCoinbase: freezed == acceptCoinbase ? _self.acceptCoinbase : acceptCoinbase // ignore: cast_nullable_to_non_nullable
@@ -167,10 +168,10 @@ return $default(_that);case _:
 /// }
 /// ```
 
-@optionalTypeArgs TResult maybeWhen<TResult extends Object?>(TResult Function(@JsonKey(name: 'asset')  String? asset, @JsonKey(name: 'min_topoheight')  int? minTopoHeight, @JsonKey(name: 'max_topoheight')  int? maxTopoHeight, @JsonKey(name: 'min_timestamp')  int? minTimestamp, @JsonKey(name: 'max_timestamp')  int? maxTimestamp, @JsonKey(name: 'address')  String? address, @JsonKey(name: 'accept_incoming')  bool? acceptIncoming, @JsonKey(name: 'accept_outgoing')  bool? acceptOutgoing, @JsonKey(name: 'accept_coinbase')  bool? acceptCoinbase, @JsonKey(name: 'accept_burn')  bool? acceptBurn, @JsonKey(name: 'accept_blob')  bool? acceptBlob, @JsonKey(name: 'query')  dynamic query, @JsonKey(name: 'limit')  int? limit, @JsonKey(name: 'skip')  int? skip)?  $default,{required TResult orElse(),}) {final _that = this;
+@optionalTypeArgs TResult maybeWhen<TResult extends Object?>(TResult Function(@JsonKey(name: 'asset')  String? asset, @JsonKey(name: 'min_topoheight')  int? minTopoHeight, @JsonKey(name: 'max_topoheight')  int? maxTopoHeight, @JsonKey(name: 'min_timestamp')  int? minTimestamp, @JsonKey(name: 'max_timestamp')  int? maxTimestamp, @JsonKey(name: 'address')  String? address, @JsonKey(name: 'contract')  String? contract, @JsonKey(name: 'accept_incoming')  bool? acceptIncoming, @JsonKey(name: 'accept_outgoing')  bool? acceptOutgoing, @JsonKey(name: 'accept_coinbase')  bool? acceptCoinbase, @JsonKey(name: 'accept_burn')  bool? acceptBurn, @JsonKey(name: 'accept_blob')  bool? acceptBlob, @JsonKey(name: 'query')  dynamic query, @JsonKey(name: 'limit')  int? limit, @JsonKey(name: 'skip')  int? skip)?  $default,{required TResult orElse(),}) {final _that = this;
 switch (_that) {
 case _ListTransactionsParams() when $default != null:
-return $default(_that.asset,_that.minTopoHeight,_that.maxTopoHeight,_that.minTimestamp,_that.maxTimestamp,_that.address,_that.acceptIncoming,_that.acceptOutgoing,_that.acceptCoinbase,_that.acceptBurn,_that.acceptBlob,_that.query,_that.limit,_that.skip);case _:
+return $default(_that.asset,_that.minTopoHeight,_that.maxTopoHeight,_that.minTimestamp,_that.maxTimestamp,_that.address,_that.contract,_that.acceptIncoming,_that.acceptOutgoing,_that.acceptCoinbase,_that.acceptBurn,_that.acceptBlob,_that.query,_that.limit,_that.skip);case _:
   return orElse();
 
 }
@@ -188,10 +189,10 @@ return $default(_that.asset,_that.minTopoHeight,_that.maxTopoHeight,_that.minTim
 /// }
 /// ```
 
-@optionalTypeArgs TResult when<TResult extends Object?>(TResult Function(@JsonKey(name: 'asset')  String? asset, @JsonKey(name: 'min_topoheight')  int? minTopoHeight, @JsonKey(name: 'max_topoheight')  int? maxTopoHeight, @JsonKey(name: 'min_timestamp')  int? minTimestamp, @JsonKey(name: 'max_timestamp')  int? maxTimestamp, @JsonKey(name: 'address')  String? address, @JsonKey(name: 'accept_incoming')  bool? acceptIncoming, @JsonKey(name: 'accept_outgoing')  bool? acceptOutgoing, @JsonKey(name: 'accept_coinbase')  bool? acceptCoinbase, @JsonKey(name: 'accept_burn')  bool? acceptBurn, @JsonKey(name: 'accept_blob')  bool? acceptBlob, @JsonKey(name: 'query')  dynamic query, @JsonKey(name: 'limit')  int? limit, @JsonKey(name: 'skip')  int? skip)  $default,) {final _that = this;
+@optionalTypeArgs TResult when<TResult extends Object?>(TResult Function(@JsonKey(name: 'asset')  String? asset, @JsonKey(name: 'min_topoheight')  int? minTopoHeight, @JsonKey(name: 'max_topoheight')  int? maxTopoHeight, @JsonKey(name: 'min_timestamp')  int? minTimestamp, @JsonKey(name: 'max_timestamp')  int? maxTimestamp, @JsonKey(name: 'address')  String? address, @JsonKey(name: 'contract')  String? contract, @JsonKey(name: 'accept_incoming')  bool? acceptIncoming, @JsonKey(name: 'accept_outgoing')  bool? acceptOutgoing, @JsonKey(name: 'accept_coinbase')  bool? acceptCoinbase, @JsonKey(name: 'accept_burn')  bool? acceptBurn, @JsonKey(name: 'accept_blob')  bool? acceptBlob, @JsonKey(name: 'query')  dynamic query, @JsonKey(name: 'limit')  int? limit, @JsonKey(name: 'skip')  int? skip)  $default,) {final _that = this;
 switch (_that) {
 case _ListTransactionsParams():
-return $default(_that.asset,_that.minTopoHeight,_that.maxTopoHeight,_that.minTimestamp,_that.maxTimestamp,_that.address,_that.acceptIncoming,_that.acceptOutgoing,_that.acceptCoinbase,_that.acceptBurn,_that.acceptBlob,_that.query,_that.limit,_that.skip);case _:
+return $default(_that.asset,_that.minTopoHeight,_that.maxTopoHeight,_that.minTimestamp,_that.maxTimestamp,_that.address,_that.contract,_that.acceptIncoming,_that.acceptOutgoing,_that.acceptCoinbase,_that.acceptBurn,_that.acceptBlob,_that.query,_that.limit,_that.skip);case _:
   throw StateError('Unexpected subclass');
 
 }
@@ -208,10 +209,10 @@ return $default(_that.asset,_that.minTopoHeight,_that.maxTopoHeight,_that.minTim
 /// }
 /// ```
 
-@optionalTypeArgs TResult? whenOrNull<TResult extends Object?>(TResult? Function(@JsonKey(name: 'asset')  String? asset, @JsonKey(name: 'min_topoheight')  int? minTopoHeight, @JsonKey(name: 'max_topoheight')  int? maxTopoHeight, @JsonKey(name: 'min_timestamp')  int? minTimestamp, @JsonKey(name: 'max_timestamp')  int? maxTimestamp, @JsonKey(name: 'address')  String? address, @JsonKey(name: 'accept_incoming')  bool? acceptIncoming, @JsonKey(name: 'accept_outgoing')  bool? acceptOutgoing, @JsonKey(name: 'accept_coinbase')  bool? acceptCoinbase, @JsonKey(name: 'accept_burn')  bool? acceptBurn, @JsonKey(name: 'accept_blob')  bool? acceptBlob, @JsonKey(name: 'query')  dynamic query, @JsonKey(name: 'limit')  int? limit, @JsonKey(name: 'skip')  int? skip)?  $default,) {final _that = this;
+@optionalTypeArgs TResult? whenOrNull<TResult extends Object?>(TResult? Function(@JsonKey(name: 'asset')  String? asset, @JsonKey(name: 'min_topoheight')  int? minTopoHeight, @JsonKey(name: 'max_topoheight')  int? maxTopoHeight, @JsonKey(name: 'min_timestamp')  int? minTimestamp, @JsonKey(name: 'max_timestamp')  int? maxTimestamp, @JsonKey(name: 'address')  String? address, @JsonKey(name: 'contract')  String? contract, @JsonKey(name: 'accept_incoming')  bool? acceptIncoming, @JsonKey(name: 'accept_outgoing')  bool? acceptOutgoing, @JsonKey(name: 'accept_coinbase')  bool? acceptCoinbase, @JsonKey(name: 'accept_burn')  bool? acceptBurn, @JsonKey(name: 'accept_blob')  bool? acceptBlob, @JsonKey(name: 'query')  dynamic query, @JsonKey(name: 'limit')  int? limit, @JsonKey(name: 'skip')  int? skip)?  $default,) {final _that = this;
 switch (_that) {
 case _ListTransactionsParams() when $default != null:
-return $default(_that.asset,_that.minTopoHeight,_that.maxTopoHeight,_that.minTimestamp,_that.maxTimestamp,_that.address,_that.acceptIncoming,_that.acceptOutgoing,_that.acceptCoinbase,_that.acceptBurn,_that.acceptBlob,_that.query,_that.limit,_that.skip);case _:
+return $default(_that.asset,_that.minTopoHeight,_that.maxTopoHeight,_that.minTimestamp,_that.maxTimestamp,_that.address,_that.contract,_that.acceptIncoming,_that.acceptOutgoing,_that.acceptCoinbase,_that.acceptBurn,_that.acceptBlob,_that.query,_that.limit,_that.skip);case _:
   return null;
 
 }
@@ -223,7 +224,7 @@ return $default(_that.asset,_that.minTopoHeight,_that.maxTopoHeight,_that.minTim
 
 @JsonSerializable(includeIfNull: false)
 class _ListTransactionsParams implements ListTransactionsParams {
-  const _ListTransactionsParams({@JsonKey(name: 'asset') this.asset, @JsonKey(name: 'min_topoheight') this.minTopoHeight, @JsonKey(name: 'max_topoheight') this.maxTopoHeight, @JsonKey(name: 'min_timestamp') this.minTimestamp, @JsonKey(name: 'max_timestamp') this.maxTimestamp, @JsonKey(name: 'address') this.address, @JsonKey(name: 'accept_incoming') this.acceptIncoming, @JsonKey(name: 'accept_outgoing') this.acceptOutgoing, @JsonKey(name: 'accept_coinbase') this.acceptCoinbase, @JsonKey(name: 'accept_burn') this.acceptBurn, @JsonKey(name: 'accept_blob') this.acceptBlob, @JsonKey(name: 'query') this.query, @JsonKey(name: 'limit') this.limit, @JsonKey(name: 'skip') this.skip});
+  const _ListTransactionsParams({@JsonKey(name: 'asset') this.asset, @JsonKey(name: 'min_topoheight') this.minTopoHeight, @JsonKey(name: 'max_topoheight') this.maxTopoHeight, @JsonKey(name: 'min_timestamp') this.minTimestamp, @JsonKey(name: 'max_timestamp') this.maxTimestamp, @JsonKey(name: 'address') this.address, @JsonKey(name: 'contract') this.contract, @JsonKey(name: 'accept_incoming') this.acceptIncoming, @JsonKey(name: 'accept_outgoing') this.acceptOutgoing, @JsonKey(name: 'accept_coinbase') this.acceptCoinbase, @JsonKey(name: 'accept_burn') this.acceptBurn, @JsonKey(name: 'accept_blob') this.acceptBlob, @JsonKey(name: 'query') this.query, @JsonKey(name: 'limit') this.limit, @JsonKey(name: 'skip') this.skip});
   factory _ListTransactionsParams.fromJson(Map<String, dynamic> json) => _$ListTransactionsParamsFromJson(json);
 
 @override@JsonKey(name: 'asset') final  String? asset;
@@ -232,6 +233,7 @@ class _ListTransactionsParams implements ListTransactionsParams {
 @override@JsonKey(name: 'min_timestamp') final  int? minTimestamp;
 @override@JsonKey(name: 'max_timestamp') final  int? maxTimestamp;
 @override@JsonKey(name: 'address') final  String? address;
+@override@JsonKey(name: 'contract') final  String? contract;
 @override@JsonKey(name: 'accept_incoming') final  bool? acceptIncoming;
 @override@JsonKey(name: 'accept_outgoing') final  bool? acceptOutgoing;
 @override@JsonKey(name: 'accept_coinbase') final  bool? acceptCoinbase;
@@ -255,16 +257,16 @@ Map<String, dynamic> toJson() {
 
 @override
 bool operator ==(Object other) {
-  return identical(this, other) || (other.runtimeType == runtimeType&&other is _ListTransactionsParams&&(identical(other.asset, asset) || other.asset == asset)&&(identical(other.minTopoHeight, minTopoHeight) || other.minTopoHeight == minTopoHeight)&&(identical(other.maxTopoHeight, maxTopoHeight) || other.maxTopoHeight == maxTopoHeight)&&(identical(other.minTimestamp, minTimestamp) || other.minTimestamp == minTimestamp)&&(identical(other.maxTimestamp, maxTimestamp) || other.maxTimestamp == maxTimestamp)&&(identical(other.address, address) || other.address == address)&&(identical(other.acceptIncoming, acceptIncoming) || other.acceptIncoming == acceptIncoming)&&(identical(other.acceptOutgoing, acceptOutgoing) || other.acceptOutgoing == acceptOutgoing)&&(identical(other.acceptCoinbase, acceptCoinbase) || other.acceptCoinbase == acceptCoinbase)&&(identical(other.acceptBurn, acceptBurn) || other.acceptBurn == acceptBurn)&&(identical(other.acceptBlob, acceptBlob) || other.acceptBlob == acceptBlob)&&const DeepCollectionEquality().equals(other.query, query)&&(identical(other.limit, limit) || other.limit == limit)&&(identical(other.skip, skip) || other.skip == skip));
+  return identical(this, other) || (other.runtimeType == runtimeType&&other is _ListTransactionsParams&&(identical(other.asset, asset) || other.asset == asset)&&(identical(other.minTopoHeight, minTopoHeight) || other.minTopoHeight == minTopoHeight)&&(identical(other.maxTopoHeight, maxTopoHeight) || other.maxTopoHeight == maxTopoHeight)&&(identical(other.minTimestamp, minTimestamp) || other.minTimestamp == minTimestamp)&&(identical(other.maxTimestamp, maxTimestamp) || other.maxTimestamp == maxTimestamp)&&(identical(other.address, address) || other.address == address)&&(identical(other.contract, contract) || other.contract == contract)&&(identical(other.acceptIncoming, acceptIncoming) || other.acceptIncoming == acceptIncoming)&&(identical(other.acceptOutgoing, acceptOutgoing) || other.acceptOutgoing == acceptOutgoing)&&(identical(other.acceptCoinbase, acceptCoinbase) || other.acceptCoinbase == acceptCoinbase)&&(identical(other.acceptBurn, acceptBurn) || other.acceptBurn == acceptBurn)&&(identical(other.acceptBlob, acceptBlob) || other.acceptBlob == acceptBlob)&&const DeepCollectionEquality().equals(other.query, query)&&(identical(other.limit, limit) || other.limit == limit)&&(identical(other.skip, skip) || other.skip == skip));
 }
 
 @JsonKey(includeFromJson: false, includeToJson: false)
 @override
-int get hashCode => Object.hash(runtimeType,asset,minTopoHeight,maxTopoHeight,minTimestamp,maxTimestamp,address,acceptIncoming,acceptOutgoing,acceptCoinbase,acceptBurn,acceptBlob,const DeepCollectionEquality().hash(query),limit,skip);
+int get hashCode => Object.hash(runtimeType,asset,minTopoHeight,maxTopoHeight,minTimestamp,maxTimestamp,address,contract,acceptIncoming,acceptOutgoing,acceptCoinbase,acceptBurn,acceptBlob,const DeepCollectionEquality().hash(query),limit,skip);
 
 @override
 String toString() {
-  return 'ListTransactionsParams(asset: $asset, minTopoHeight: $minTopoHeight, maxTopoHeight: $maxTopoHeight, minTimestamp: $minTimestamp, maxTimestamp: $maxTimestamp, address: $address, acceptIncoming: $acceptIncoming, acceptOutgoing: $acceptOutgoing, acceptCoinbase: $acceptCoinbase, acceptBurn: $acceptBurn, acceptBlob: $acceptBlob, query: $query, limit: $limit, skip: $skip)';
+  return 'ListTransactionsParams(asset: $asset, minTopoHeight: $minTopoHeight, maxTopoHeight: $maxTopoHeight, minTimestamp: $minTimestamp, maxTimestamp: $maxTimestamp, address: $address, contract: $contract, acceptIncoming: $acceptIncoming, acceptOutgoing: $acceptOutgoing, acceptCoinbase: $acceptCoinbase, acceptBurn: $acceptBurn, acceptBlob: $acceptBlob, query: $query, limit: $limit, skip: $skip)';
 }
 
 
@@ -275,7 +277,7 @@ abstract mixin class _$ListTransactionsParamsCopyWith<$Res> implements $ListTran
   factory _$ListTransactionsParamsCopyWith(_ListTransactionsParams value, $Res Function(_ListTransactionsParams) _then) = __$ListTransactionsParamsCopyWithImpl;
 @override @useResult
 $Res call({
-@JsonKey(name: 'asset') String? asset,@JsonKey(name: 'min_topoheight') int? minTopoHeight,@JsonKey(name: 'max_topoheight') int? maxTopoHeight,@JsonKey(name: 'min_timestamp') int? minTimestamp,@JsonKey(name: 'max_timestamp') int? maxTimestamp,@JsonKey(name: 'address') String? address,@JsonKey(name: 'accept_incoming') bool? acceptIncoming,@JsonKey(name: 'accept_outgoing') bool? acceptOutgoing,@JsonKey(name: 'accept_coinbase') bool? acceptCoinbase,@JsonKey(name: 'accept_burn') bool? acceptBurn,@JsonKey(name: 'accept_blob') bool? acceptBlob,@JsonKey(name: 'query') dynamic query,@JsonKey(name: 'limit') int? limit,@JsonKey(name: 'skip') int? skip
+@JsonKey(name: 'asset') String? asset,@JsonKey(name: 'min_topoheight') int? minTopoHeight,@JsonKey(name: 'max_topoheight') int? maxTopoHeight,@JsonKey(name: 'min_timestamp') int? minTimestamp,@JsonKey(name: 'max_timestamp') int? maxTimestamp,@JsonKey(name: 'address') String? address,@JsonKey(name: 'contract') String? contract,@JsonKey(name: 'accept_incoming') bool? acceptIncoming,@JsonKey(name: 'accept_outgoing') bool? acceptOutgoing,@JsonKey(name: 'accept_coinbase') bool? acceptCoinbase,@JsonKey(name: 'accept_burn') bool? acceptBurn,@JsonKey(name: 'accept_blob') bool? acceptBlob,@JsonKey(name: 'query') dynamic query,@JsonKey(name: 'limit') int? limit,@JsonKey(name: 'skip') int? skip
 });
 
 
@@ -292,7 +294,7 @@ class __$ListTransactionsParamsCopyWithImpl<$Res>
 
 /// Create a copy of ListTransactionsParams
 /// with the given fields replaced by the non-null parameter values.
-@override @pragma('vm:prefer-inline') $Res call({Object? asset = freezed,Object? minTopoHeight = freezed,Object? maxTopoHeight = freezed,Object? minTimestamp = freezed,Object? maxTimestamp = freezed,Object? address = freezed,Object? acceptIncoming = freezed,Object? acceptOutgoing = freezed,Object? acceptCoinbase = freezed,Object? acceptBurn = freezed,Object? acceptBlob = freezed,Object? query = freezed,Object? limit = freezed,Object? skip = freezed,}) {
+@override @pragma('vm:prefer-inline') $Res call({Object? asset = freezed,Object? minTopoHeight = freezed,Object? maxTopoHeight = freezed,Object? minTimestamp = freezed,Object? maxTimestamp = freezed,Object? address = freezed,Object? contract = freezed,Object? acceptIncoming = freezed,Object? acceptOutgoing = freezed,Object? acceptCoinbase = freezed,Object? acceptBurn = freezed,Object? acceptBlob = freezed,Object? query = freezed,Object? limit = freezed,Object? skip = freezed,}) {
   return _then(_ListTransactionsParams(
 asset: freezed == asset ? _self.asset : asset // ignore: cast_nullable_to_non_nullable
 as String?,minTopoHeight: freezed == minTopoHeight ? _self.minTopoHeight : minTopoHeight // ignore: cast_nullable_to_non_nullable
@@ -300,6 +302,7 @@ as int?,maxTopoHeight: freezed == maxTopoHeight ? _self.maxTopoHeight : maxTopoH
 as int?,minTimestamp: freezed == minTimestamp ? _self.minTimestamp : minTimestamp // ignore: cast_nullable_to_non_nullable
 as int?,maxTimestamp: freezed == maxTimestamp ? _self.maxTimestamp : maxTimestamp // ignore: cast_nullable_to_non_nullable
 as int?,address: freezed == address ? _self.address : address // ignore: cast_nullable_to_non_nullable
+as String?,contract: freezed == contract ? _self.contract : contract // ignore: cast_nullable_to_non_nullable
 as String?,acceptIncoming: freezed == acceptIncoming ? _self.acceptIncoming : acceptIncoming // ignore: cast_nullable_to_non_nullable
 as bool?,acceptOutgoing: freezed == acceptOutgoing ? _self.acceptOutgoing : acceptOutgoing // ignore: cast_nullable_to_non_nullable
 as bool?,acceptCoinbase: freezed == acceptCoinbase ? _self.acceptCoinbase : acceptCoinbase // ignore: cast_nullable_to_non_nullable